+++ src/main/java/mod/gottsch/forge/magic_treasures/core/loot/modifier/LootModifierByRarity.java
@@ -35,12 +35,14 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
+import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.common.loot.IGlobalLootModifier;
 import net.minecraftforge.common.loot.LootModifier;
 import net.minecraftforge.registries.RegistryObject;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Supplier;
 
 /**
@@ -81,7 +83,15 @@ public class LootModifierByRarity extends LootModifier {
 		// if chance was left blank or null, then set to 100% by default
 		double localChance = chance == 0.0 ? 1.0 : chance;
 
-		if (Config.SERVER.loot.enableVanillaLootModifiers.get() && RandomHelper.checkProbability(context.getLevel().getRandom(), localChance * 100)) {
+		// determine if specific loot modifier is enabled
+		boolean isEnabled = Optional.ofNullable(Config.enableLootModifiers.get(rarity.toLowerCase())).
+				map(ForgeConfigSpec.ConfigValue::get).orElse(false);
+		MagicTreasures.LOGGER.debug("isEnabled for {} -> {}", rarity, isEnabled);
+
+		if (Config.SERVER.loot.enableVanillaLootModifiers.get()
+				&& isEnabled
+				&& RandomHelper.checkProbability(context.getLevel().getRandom(), localChance * 100)) {
+
 			IRarity rarity = MagicTreasuresApi.getRarity(this.rarity).orElse(MagicTreasuresRarity.NONE);
 			List<Item> lootList = JewelryRegistry.get(rarity);
 			lootList.addAll(StoneRegistry.get(rarity));