+++ src/main/java/mod/gottsch/forge/magic_treasures/datagen/LanguageGen.java
@@ -47,6 +47,7 @@ public class LanguageGen extends LanguageProvider {
 
         // metals
         add(MagicTreasuresItems.SILVER_INGOT.get(), "Silver Ingot");
+        add(MagicTreasuresItems.RAW_SILVER.get(), "Raw Silver");
 
         // stones
         add(MagicTreasuresItems.TOPAZ.get(), "Topaz");