+++ src/main/java/mod/gottsch/forge/magic_treasures/datagen/MagicTreasuresLootTableProvider.java
@@ -1,0 +1,39 @@
+package mod.gottsch.forge.magic_treasures.datagen;
+
+import com.google.common.collect.ImmutableList;
+import com.mojang.datafixers.util.Pair;
+import net.minecraft.data.DataGenerator;
+import net.minecraft.data.loot.LootTableProvider;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraft.world.level.storage.loot.LootTables;
+import net.minecraft.world.level.storage.loot.ValidationContext;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParamSet;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+import java.util.Map;
+import java.util.function.BiConsumer;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
+
+/**
+ * Created by Mark Gottschling on Nov 12, 2024
+ */
+public class MagicTreasuresLootTableProvider extends LootTableProvider {
+
+    public MagicTreasuresLootTableProvider(DataGenerator dataGenerator) {
+        super(dataGenerator);
+    }
+
+    @Override
+    protected @NotNull List<Pair<Supplier<Consumer<BiConsumer<ResourceLocation, LootTable.Builder>>>, LootContextParamSet>> getTables() {
+        return ImmutableList.of(Pair.of(BlockLootTables::new, LootContextParamSets.BLOCK));
+    }
+
+    @Override
+    protected void validate(Map<ResourceLocation, LootTable> map, ValidationContext validationtracker) {
+        map.forEach((location, lootTable) -> LootTables.validate(validationtracker, location, lootTable));
+    }
+}