+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-11-12
+
+### Changed
+- applied Config options to control each Loot Modifier for LootModifierByRarity
+- fix Raw Silver smelting recipe
+- remove guide on first join. Still available in Creative tab.
+
+### Added
+- Russian lang by CoolDemon96.
+- Raw Silver blast furnace recipe.
+
+## [1.2.0] - 2024-11-12
+
+### Changed
+- Silver Ore blocks now drop Raw Silver items.
+- Raw Silver can be smelted into Silver Ingots.
+- Use datagen for blockstate, block model and block loot tables.
+
+### Added
+- Raw Silver Item
+- Raw Silver to the Forge tags (aka ore dictionary)
+- Config options to control each Loot Modifier
+
 ## [1.1.0] - 2024-07-15
 
 ### Changed